Adds mosquito
I'm not sure if this is the right places in the list, but I've added [my mosquito project](https://github.com/robacarp/mosquito/) to the Queue section. It also includes jobs which are scheduled (eg every X minutes), so I considered putting it in the `Scheduling` section, but it is intended firstly as a replacement for sidekiq so I put it alongside.