Adding fast-crystal project
Not sure if it belongs in "resources" or "funny stuff". But, I found it educational, coming from Ruby, to port most of the benchmarks in fast-ruby, as well as figure out why some of the Ruby idioms won't work in Crystal. Hopefully, it's going to be a good *resource* for people going through the same process.