Add Keimeno TUI engine
I have no idea if the project is mature enough to warrant being on here. I use this library for a handful of my own utilities and it's been great. I don't have tests written for it. I just think it'd be good to get some other eyes on it.

I also don't know if it should be in CLI Builders or CLI Utilities. The list of builders is _essentially_ just option parsers, but utilities are ... applications which run in the console.